+++ bundles/framework/layerlist/view/LayerViewTabs/SelectedLayers/LayerBox.jsx
@@ -2,8 +2,9 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { Row, Col, ColAuto, ColAutoRight } from './Grid';
-import { Slider, Icon, NumberInput } from 'oskari-ui';
-import { EyeOpen, EyeShut, DataLayerIcon, ImageLayerIcon, DragIcon } from './CustomIcons/CustomIcons';
+import { Icon } from 'oskari-ui';
+import { EyeOpen, EyeShut, DragIcon } from './CustomIcons/CustomIcons';
+import { LayerInfoBox } from './LayerInfoBox';
 
 const StyledBox = styled.div`
     min-height: 100px;
@@ -20,34 +21,6 @@ const GrayRow = styled(Row)`
     padding-left: 60px;
 `;
 
-const StyledSlider = styled.div`
-    border: solid 2px #d9d9d9;
-    border-radius: 4px;
-    width: 150px;
-    padding: 8px 15px;
-`;
-
-const StyledNumberInput = styled(NumberInput)`
-    width: 70px !important;
-    font-size: 15px;
-    box-shadow: inset 1px 1px 4px 0 rgba(87, 87, 87, 0.26);
-`;
-
-const LayerIcon = ({ type }) => {
-    if (type === 'wmts') {
-        return (
-            <ImageLayerIcon style={{ marginTop: '5px' }} title={type} />
-        );
-    }
-    return (
-        <DataLayerIcon style={{ marginTop: '5px' }} title={type} />
-    );
-};
-
-LayerIcon.propTypes = {
-    type: PropTypes.string
-};
-
 export const LayerBox = ({ layer }) => {
     const [slider, setSlider] = useState(layer.getOpacity());
     const name = layer.getName();
@@ -59,24 +32,19 @@ export const LayerBox = ({ layer }) => {
     // Try to find this somewhere
     // const activeFeats = layer.getActiveFeatures().length;
     const handleOpacityChange = value => {
-        // TODO send event instead of manipulating layer
         layer.setOpacity(value);
         setSlider(value);
     };
     const handleDragEvent = () => {
-        // TODO
         console.log('Drag');
     };
     const handleToggleVisibility = () => {
-        // TODO
         console.log('Toggle visibility');
     };
     const handleOpenMenu = () => {
-        // TODO
         console.log('Open menu');
     };
     const handleRemoveLayer = () => {
-        // TODO
         console.log('Remove layer');
     };
     return (
@@ -101,34 +69,12 @@ export const LayerBox = ({ layer }) => {
                         </ColAutoRight>
                     </Row>
                     <GrayRow>
-                        <ColAuto>
-                            <LayerIcon type={layerType} />
-                        </ColAuto>
-                        <ColAuto>
-                            <StyledSlider>
-                                <Slider
-                                    value={slider}
-                                    onChange={handleOpacityChange}
-                                    style={{ margin: '0px' }}
-                                />
-                            </StyledSlider>
-                        </ColAuto>
-                        <ColAuto>
-                            <StyledNumberInput
-                                min={0}
-                                max={100}
-                                value={slider}
-                                onChange={handleOpacityChange}
-                                formatter={value => `${value} %`}
-                            />
-                        </ColAuto>
-                        <ColAutoRight>
-                            <Icon
-                                type="menu"
-                                onClick={handleOpenMenu}
-                                style={{ color: '#006ce8', fontSize: '16px', marginTop: '8px' }}
-                            />
-                        </ColAutoRight>
+                        <LayerInfoBox
+                            layerType={layerType}
+                            slider={slider}
+                            handleOpacityChange={handleOpacityChange}
+                            handleOpenMenu={handleOpenMenu}
+                        />
                     </GrayRow>
                 </StyledBox>
             </Col>
